refactor: extract PD load sharing to power_distribution.py
Move the multi-battery load-sharing cluster out of the
ChargeDischargeController monolith into power_distribution.py:
  _distribute_power_by_limits, _select_batteries_for_operation,
  _rebalance_expired_load_sharing_hold, and the _round_to_5w helper.

The wall-clock split-load holds (_charge/_discharge_selection_hold_until)
move with the cluster into PowerDistribution. The active-battery lists
stay on the controller because sensor.py, switch.py, and the control loop
read/mutate them; the module accesses them via self._controller. The
per-battery limit and capacity primitives (_battery_power_limit,
_clamp_to_system_capacity) also stay on the controller and are queried as
inputs.

PowerDistribution is instantiated in async_setup_entry next to
ExternalLoads. Call sites updated to self._power_distribution.*. The
characterization tests added in the previous commit retargeted to the
module with identical assertions and stay green = cero cambio funcional.
